Create shared profile button config method for reuse
